move HumanString method from task/elbv2 package to describe package
