+++ internal/pkg/aws/ecs/task_test.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/ecs"
-	"github.com/dustin/go-humanize"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/require"
 )
@@ -307,146 +306,6 @@ func TestTask_PrivateIP(t *testing.T) {
 	}
 }
 
-func TestTaskStatus_HumanString(t *testing.T) {
-	// from the function changes (ex: from "1 month ago" to "2 months ago"). To make our tests stable,
-	oldHumanize := humanizeTime
-	humanizeTime = func(then time.Time) string {
-		now, _ := time.Parse(time.RFC3339, "2020-01-01T00:00:00+00:00")
-		return humanize.RelTime(then, now, "ago", "from now")
-	}
-	defer func() {
-		humanizeTime = oldHumanize
-	}()
-	startTime, _ := time.Parse(time.RFC3339, "2006-01-02T15:04:05+00:00")
-	stopTime, _ := time.Parse(time.RFC3339, "2006-01-02T16:04:05+00:00")
-	mockImageDigest := "18f7eb6cff6e63e5f5273fb53f672975fe6044580f66c354f55d2de8dd28aec7"
-	testCases := map[string]struct {
-		id               string
-		health           string
-		lastStatus       string
-		imageDigest      string
-		startedAt        time.Time
-		stoppedAt        time.Time
-		capacityProvider string
-
-		wantTaskStatus string
-	}{
-		"all params": {
-			health:           "HEALTHY",
-			id:               "aslhfnqo39j8qomimvoiqm89349",
-			lastStatus:       "RUNNING",
-			startedAt:        startTime,
-			stoppedAt:        stopTime,
-			imageDigest:      mockImageDigest,
-			capacityProvider: "FARGATE",
-
-			wantTaskStatus: "aslhfnqo\t18f7eb6c\tRUNNING\t14 years ago\tFARGATE\tHEALTHY",
-		},
-		"missing params": {
-			health:     "HEALTHY",
-			lastStatus: "RUNNING",
-
-			wantTaskStatus: "-\t-\tRUNNING\t-\t-\tHEALTHY",
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			// GIVEN
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-
-			task := TaskStatus{
-				Health: tc.health,
-				ID:     tc.id,
-				Images: []Image{
-					{
-						Digest: tc.imageDigest,
-					},
-				},
-				LastStatus:       tc.lastStatus,
-				StartedAt:        tc.startedAt,
-				StoppedAt:        tc.stoppedAt,
-				CapacityProvider: tc.capacityProvider,
-			}
-
-			gotTaskStatus := task.HumanString()
-
-			require.Equal(t, tc.wantTaskStatus, gotTaskStatus)
-		})
-
-	}
-}
-
-func TestStoppedTaskStatus_HumanString(t *testing.T) {
-	// from the function changes (ex: from "1 month ago" to "2 months ago"). To make our tests stable,
-	oldHumanize := humanizeTime
-	humanizeTime = func(then time.Time) string {
-		now, _ := time.Parse(time.RFC3339, "2020-01-01T00:00:00+00:00")
-		return humanize.RelTime(then, now, "ago", "from now")
-	}
-	defer func() {
-		humanizeTime = oldHumanize
-	}()
-	startTime, _ := time.Parse(time.RFC3339, "2006-01-02T15:04:05+00:00")
-	stopTime, _ := time.Parse(time.RFC3339, "2006-01-02T16:04:05+00:00")
-	mockImageDigest := "18f7eb6cff6e63e5f5273fb53f672975fe6044580f66c354f55d2de8dd28aec7"
-	testCases := map[string]struct {
-		id            string
-		health        string
-		lastStatus    string
-		imageDigest   string
-		startedAt     time.Time
-		stoppedAt     time.Time
-		stoppedReason string
-
-		wantTaskStatus string
-	}{
-		"all params": {
-			health:        "HEALTHY",
-			id:            "aslhfnqo39j8qomimvoiqm89349",
-			lastStatus:    "RUNNING",
-			startedAt:     startTime,
-			stoppedAt:     stopTime,
-			imageDigest:   mockImageDigest,
-			stoppedReason: "Stopped by reasons",
-
-			wantTaskStatus: "aslhfnqo\t18f7eb6c\tRUNNING\t14 years ago\t14 years ago\tStopped by reasons",
-		},
-		"missing params": {
-			lastStatus: "RUNNING",
-
-			wantTaskStatus: "-\t-\tRUNNING\t-\t-\t-",
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			// GIVEN
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-
-			task := TaskStatus{
-				ID: tc.id,
-				Images: []Image{
-					{
-						Digest: tc.imageDigest,
-					},
-				},
-				LastStatus:    tc.lastStatus,
-				StartedAt:     tc.startedAt,
-				StoppedAt:     tc.stoppedAt,
-				StoppedReason: tc.stoppedReason,
-			}
-
-			gotTaskStatus := (StoppedTaskStatus)(task).HumanString()
-
-			require.Equal(t, tc.wantTaskStatus, gotTaskStatus)
-		})
-
-	}
-}
-
 func Test_TaskID(t *testing.T) {
 	testCases := map[string]struct {
 		taskARN string
@@ -495,7 +354,7 @@ func Test_ShortTaskID(t *testing.T) {
 	for name, tc := range testCases {
 		t.Run(name, func(t *testing.T) {
 			// WHEN
-			gotID := ShortTaskID(tc.taskID)
+			gotID := shortTaskID(tc.taskID)
 
 			// THEN
 			require.Equal(t, tc.wantID, gotID)
