+++ internal/pkg/aws/elbv2/elbv2_test.go
@@ -137,44 +137,3 @@ func TestELBV2_TargetsHealth(t *testing.T) {
 		})
 	}
 }
-
-func TestTargetHealth_HumanString(t *testing.T) {
-	testCases := map[string]struct {
-		id     string
-		state  string
-		reason string
-
-		wantedOut string
-	}{
-		"healthy": {
-			id:    "target-1",
-			state: elbv2.TargetHealthStateEnumHealthy,
-
-			wantedOut: `target-1	-	HEALTHY`,
-		},
-		"unhealthy with reason": {
-			id:     "target-1",
-			state:  elbv2.TargetHealthStateEnumUnhealthy,
-			reason: elbv2.TargetHealthReasonEnumTargetResponseCodeMismatch,
-
-			wantedOut: `target-1	Target.ResponseCodeMismatch	UNHEALTHY`,
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			th := TargetHealth{
-				Target: &elbv2.TargetDescription{
-					Id: aws.String(tc.id),
-				},
-				TargetHealth: &elbv2.TargetHealth{
-					State:  aws.String(tc.state),
-					Reason: aws.String(tc.reason),
-				},
-			}
-
-			got := th.HumanString()
-			require.Equal(t, tc.wantedOut, got)
-		})
-	}
-}
